+++ src/main/groovy/info/solidsoft/gradle/pitest/PitestPluginExtension.groovy
@@ -17,8 +17,12 @@ package info.solidsoft.gradle.pitest
 
 import groovy.transform.CompileStatic
 import org.gradle.api.Incubating
+import org.gradle.api.Project
+import org.gradle.api.model.ObjectFactory
+import org.gradle.api.provider.ListProperty
+import org.gradle.api.provider.Property
+import org.gradle.api.provider.SetProperty
 import org.gradle.api.tasks.SourceSet
-import org.gradle.api.tasks.TaskInstantiationException
 
 /**
  * Extension class with configurable parameters for Pitest plugin.
@@ -29,7 +33,7 @@ import org.gradle.api.tasks.TaskInstantiationException
 @CompileStatic
 class PitestPluginExtension {
 
-    String pitestVersion
+    final Property<String> pitestVersion
 
     /**
      * Specifies what test plugin to use.
@@ -38,44 +42,59 @@ class PitestPluginExtension {
      *
      * @since 1.3.0
      */
-    String testPlugin
-    File reportDir
-    Set<String> targetClasses
-    Set<String> targetTests
-    Integer dependencyDistance
-    Integer threads
-    Boolean mutateStaticInits
-    Boolean includeJarFiles
-    Set<String> mutators
-    Set<String> excludedMethods
-    Set<String> excludedClasses
-    Set<String> avoidCallsTo
-    Boolean verbose
-    BigDecimal timeoutFactor
-    Integer timeoutConstInMillis
-    Integer maxMutationsPerClass
+    final Property<String> testPlugin
+
+//    //ClassNotFoundException: org.gradle.api.file.FileSystemLocationProperty in Gradle <5.6 due to super interface of RegularFileProperty
+//    final RegularFileProperty reportDir
+    private File reportDir
+
+    final SetProperty<String> targetClasses
+    final SetProperty<String> targetTests
+    final Property<Integer> dependencyDistance
+    final Property<Integer> threads
+    @Deprecated //PIT doesn't know it
+    final Property<Boolean> mutateStaticInits
+    @Deprecated //removed in PIT 0.33
+    final Property<Boolean> includeJarFiles
+    final SetProperty<String> mutators
+    final SetProperty<String> excludedMethods
+    final SetProperty<String> excludedClasses
+
+    /**
+     * A list of test classes which should be excluded when mutating.
+     *
+     * @since 1.3.0
+     * @see #excludedClasses
+     * @see #excludedMethods
+     */
+    @Incubating
+    final SetProperty<String> excludedTestClasses
+    final SetProperty<String> avoidCallsTo
+    final Property<Boolean> verbose
+    final Property<BigDecimal> timeoutFactor
+    final Property<Integer> timeoutConstInMillis
+    final Property<Integer> maxMutationsPerClass
     /**
      * JVM arguments to use when PIT launches child processes
      *
      * Note. This parameter type was changed from String to List<String> in 0.33.0.
      */
-    List<String> jvmArgs
-    Set<String> outputFormats
-    Boolean failWhenNoMutations
-    Set<String> includedGroups  //renamed from includedTestNGGroups in 1.0.0 - to adjust to changes in PIT
-    Set<String> excludedGroups  //renamed from excludedTestNGGroups in 1.0.0 - to adjust to changes in PIT
-//    File configFile           //removed in 1.1.6 to adjust to changes in PIT
-    Boolean detectInlinedCode   //new in PIT 0.28
-    Boolean timestampedReports
+    final ListProperty<String> jvmArgs
+    final SetProperty<String> outputFormats
+    final Property<Boolean> failWhenNoMutations
+    final SetProperty<String> includedGroups  //renamed from includedTestNGGroups in 1.0.0 - to adjust to changes in PIT
+    final SetProperty<String> excludedGroups  //renamed from excludedTestNGGroups in 1.0.0 - to adjust to changes in PIT
+    final Property<Boolean> detectInlinedCode   //new in PIT 0.28
+    final Property<Boolean> timestampedReports
     File historyInputLocation   //new in PIT 0.29
     File historyOutputLocation
-    Boolean enableDefaultIncrementalAnalysis    //specific for Gradle plugin - since 0.29.0
-    Integer mutationThreshold   //new in PIT 0.30
-    Integer coverageThreshold   //new in PIT 0.32
-    String mutationEngine
-    Set<SourceSet> testSourceSets   //specific for Gradle plugin - since 0.30.1
-    Set<SourceSet> mainSourceSets   //specific for Gradle plugin - since 0.30.1
-    Boolean exportLineCoverage  //new in PIT 0.32 - for debugging usage only
+    final Property<Boolean> enableDefaultIncrementalAnalysis    //specific for Gradle plugin - since 0.29.0
+    final Property<Integer> mutationThreshold   //new in PIT 0.30
+    final Property<Integer> coverageThreshold   //new in PIT 0.32
+    final Property<String> mutationEngine
+    final SetProperty<SourceSet> testSourceSets   //specific for Gradle plugin - since 0.30.1
+    final SetProperty<SourceSet> mainSourceSets   //specific for Gradle plugin - since 0.30.1
+    final Property<Boolean> exportLineCoverage  //new in PIT 0.32 - for debugging usage only
     File jvmPath    //new in PIT 0.32
 
     /**
@@ -83,7 +102,7 @@ class PitestPluginExtension {
      *
      * @since 0.33.0 (specific for Gradle plugin)
      */
-    List<String> mainProcessJvmArgs
+    final ListProperty<String> mainProcessJvmArgs
 
     /**
      * Additional mutableCodePaths (paths with production classes which should be mutated).<p/>
@@ -127,7 +146,7 @@ class PitestPluginExtension {
      */
     Map<String, String> pluginConfiguration
 
-    Integer maxSurviving    //new in PIT 1.1.10
+    final Property<Integer> maxSurviving    //new in PIT 1.1.10
 
     /**
      * Use classpath file instead of passing classpath in a command line
@@ -138,7 +157,7 @@ class PitestPluginExtension {
      * @since 1.2.0
      */
     @Incubating
-    boolean useClasspathFile = false
+    final Property<Boolean> useClasspathFile
 
     /**
      * Turnes on/off features in PIT itself and its plugins.
@@ -148,7 +167,7 @@ class PitestPluginExtension {
      * @since 1.2.1
      */
     @Incubating
-    List<String> features
+    final ListProperty<String> features
 
     /**
      * File extensions which should be filtered from a classpath.
@@ -165,37 +184,75 @@ class PitestPluginExtension {
      *
      * This feature is specific to the Gradle plugin.
      *
+     * <b>Please note</b>. Starting with 1.4.5 due to Gradle limitations only the new syntax with addFileExtensionsToFilter(['xml'[) is available.
+     * More information: https://github.com/gradle/gradle/issues/10475
+     *
      * @since 1.2.4
      */
     @Incubating
-    List<String> fileExtensionsToFilter
+    final ListProperty<String> fileExtensionsToFilter
 
-    /**
-     * A list of test classes which should be excluded when mutating.
-     *
-     * @since 1.3.0
-     * @see #excludedClasses
-     * @see #excludedMethods
-     */
-    @Incubating
-    Set<String> excludedTestClasses
+    PitestPluginExtension(Project project) {
+        ObjectFactory of = project.objects
+        pitestVersion = of.property(String)
+
+        testPlugin = of.property(String)
+//        reportDir = of.fileProperty()
+        targetClasses = setSetPropertyToNull(of.setProperty(String))    //empty collection (default value) is resolved as set to target classes
+        targetTests = of.setProperty(String)
+        dependencyDistance = setPropertyToNull(of.property(Integer))
+        threads = setPropertyToNull(of.property(Integer))
+        mutateStaticInits = setPropertyToNull(of.property(Boolean))
+        includeJarFiles = setPropertyToNull(of.property(Boolean))
+        mutators = of.setProperty(String)
+        excludedMethods = of.setProperty(String)
+        excludedClasses = of.setProperty(String)
+        excludedTestClasses = of.setProperty(String)
+        avoidCallsTo = of.setProperty(String)
+        verbose = setPropertyToNull(of.property(Boolean))
+        timeoutFactor = of.property(BigDecimal)
+        timeoutConstInMillis = setPropertyToNull(of.property(Integer))
+        maxMutationsPerClass = setPropertyToNull(of.property(Integer))
+
+        jvmArgs = setListPropertyToNull(of.listProperty(String))
+        outputFormats = of.setProperty(String)
+        failWhenNoMutations = setPropertyToNull(of.property(Boolean))
+        includedGroups = of.setProperty(String)
+        excludedGroups = of.setProperty(String)
+        detectInlinedCode = setPropertyToNull(of.property(Boolean))
+        timestampedReports = setPropertyToNull(of.property(Boolean))
+//        historyInputLocation = of.fileProperty()
+//        historyOutputLocation = of.fileProperty()
+        enableDefaultIncrementalAnalysis = setPropertyToNull(of.property(Boolean))
+        mutationThreshold = setPropertyToNull(of.property(Integer))
+        coverageThreshold = setPropertyToNull(of.property(Integer))
+        mutationEngine = of.property(String)
+        testSourceSets = setSetPropertyToNull(of.setProperty(SourceSet))
+        mainSourceSets = setSetPropertyToNull(of.setProperty(SourceSet))
+        exportLineCoverage = setPropertyToNull(of.property(Boolean))
+//        jvmPath = of.fileProperty()
+
+        mainProcessJvmArgs = of.listProperty(String)
+//        additionalMutableCodePaths = setSetPropertyToNull(of.setProperty(File))
+//        pluginConfiguration = of.mapProperty(String, String)
+        maxSurviving = setPropertyToNull(of.property(Integer))
+
+        useClasspathFile = setPropertyToNull(of.property(Boolean))
+        features = of.listProperty(String)
+
+        fileExtensionsToFilter = of.listProperty(String)
+    }
+
+    void setReportDir(File reportDir) {
+        this.reportDir = reportDir
+    }
+
+    File getReportDir() {
+        return reportDir
+    }
 
     void setReportDir(String reportDirAsString) {
         this.reportDir = new File(reportDirAsString)
-    }
-
-    void setSourceDirsAsFiles(Set<File> sourceDirs) {
-        throwExceptionAboutRemovedManualSettingOfSourceDirs()
-    }
-
-    void setSourceDirs(Set<String> sourceDirs) {
-        throwExceptionAboutRemovedManualSettingOfSourceDirs()
-    }
-
-    private throwExceptionAboutRemovedManualSettingOfSourceDirs() {
-        throw new TaskInstantiationException("Manual setting of sourceDirs was removed in version 0.30.1. " +
-                "Use mainSourceSets property to select source sets which would be used to get source directories. " +
-                "Feel free to raise an issue if you need removed feature.")
     }
 
     void setHistoryInputLocation(String historyInputLocationPath) {
@@ -211,7 +268,31 @@ class PitestPluginExtension {
     }
 
     void setTimeoutFactor(String timeoutFactor) {
-        this.timeoutFactor = new BigDecimal(timeoutFactor)
+        this.timeoutFactor.set(new BigDecimal(timeoutFactor))
+    }
+
+    /**
+     * File extensions which should be filtered from a classpath (ListProperty edition).
+     *
+     * PIT fails on not Java specific file passed on a classpath (e.g. native libraries). Native libraries ('*.so', '*.dll', '*.dylib')
+     * and '*.pom' files are filtered by default, but a developer can add extra extensions to the list:
+     * <pre>
+     * pitest {
+     *     addFileExtensionsToFilter(['xml', 'orbit'])
+     * }
+     * </pre>
+     *
+     * Rationale: https://github.com/szpak/gradle-pitest-plugin/issues/53
+     *
+     * This syntax is a workaround caused by a lack of support for "+=" with ListProperty: https://github.com/gradle/gradle/issues/10475
+     *
+     * This feature is specific to the Gradle plugin.
+     *
+     * @since 1.4.5
+     */
+    @Incubating
+    void addFileExtensionsToFilter(List<String> fileExtensionsToFilterToAdd) {
+        this.fileExtensionsToFilter.set(this.fileExtensionsToFilter.get() + fileExtensionsToFilterToAdd)
     }
 
     /**
@@ -222,18 +303,24 @@ class PitestPluginExtension {
      * @since 1.1.10
      */
     void setWithHistory(Boolean withHistory) {
-        this.enableDefaultIncrementalAnalysis = withHistory
+        this.enableDefaultIncrementalAnalysis.set(withHistory)
     }
 
-    /**
-     * The first (broken) implementation of using a file to pass additional classpath to PIT.
-     * Use "useClasspathFile" property instead.
-     *
-     * @since 1.1.11
-     */
-    @Deprecated //as of 1.2.0
-    void setClassPathFile(File classPathFile) {
-        throw new TaskInstantiationException("Passing 'classPathFile' manually was broken and it is no longer available. Use 'useClasspathFile' " +
-            "property to enable passing classpath to PIT as file. ")
+    private <T> SetProperty<T> setSetPropertyToNull(SetProperty<T> setProperty) {
+        // TODO: Switch to value() once 5.6+ is minimal required Gradle version
+        setProperty.set(null as Set)
+        return setProperty
+    }
+
+    private <T> ListProperty<T> setListPropertyToNull(ListProperty<T> listProperty) {
+        //TODO: Switch to value() once 5.6+ is minimal required Gradle version
+        listProperty.set(null as List)
+        return listProperty
+    }
+
+    //Workaround on issue with Gradle <5.0 where Integer/Boolean property had 0/false provided by default - changed in 5.0+
+    private <T> Property<T> setPropertyToNull(Property<T> property) {
+        property.set(null as T)
+        return property
     }
 }