+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # gradle-pitest-plugin changelog
 
-## 1.4.5 - UNRELEASED
+## 1.4.5 - Unreleased
 
+ - Rework internal plugin implementation to Gradle 5+ standards
  - PIT 1.4.10 by default
  - Basic Java 12 support tested by CI build
  - Move `pitest` task to `verification` group - [#136](https://github.com/szpak/gradle-pitest-plugin/issues/136) - PR by [Björn Kautler](https://github.com/Vampire)
  - Remove deprecation warnings in Gradle 6.0
+ - Bump minimal Gradle version to 4.9
  - Switch build to Gradle 5.6.1
 
 **Known limitations**. This is a technical release to cope with the changes in Gradle 5 and 6. PIT 1.4.10 is used by default, but not all